+++ fixtures/Path.php
@@ -3,7 +3,10 @@ declare(strict_types = 1);
 
 namespace Fixtures\Innmind\Url;
 
-use Innmind\Url\Path as Model;
+use Innmind\Url\{
+    Path as Model,
+    Url,
+};
 use Innmind\BlackBox\Set;
 
 final class Path
@@ -13,7 +16,32 @@ final class Path
      */
     public static function any(): Set
     {
-        return self::strings()->map(Model::of(...));
+        return Set::either(
+            self::relative(),
+            self::absolute(),
+        );
+    }
+
+    /**
+     * @return Set<Model>
+     */
+    public static function relative(): Set
+    {
+        return self::strings()
+            ->map(static fn($value) => \ltrim($value, '/'))
+            ->exclude(static fn($value) => \str_ends_with($value, '\\'))
+            ->map(Model::of(...));
+    }
+
+    /**
+     * @return Set<Model>
+     */
+    public static function absolute(): Set
+    {
+        return self::strings()
+            ->map(static fn($value) => '/'.\ltrim($value, '/'))
+            ->exclude(static fn($value) => \str_ends_with($value, '\\'))
+            ->map(Model::of(...));
     }
 
     /**
@@ -32,8 +60,27 @@ final class Path
     private static function strings(): Set
     {
         return Set::strings()
+            ->madeOf(
+                Set::strings()
+                    ->chars()
+                    ->ascii()
+                    ->exclude(static fn($char) => $char === '?')
+                    ->exclude(static fn($char) => $char === '#'),
+                Set::either(
+                    Set::strings()->unicode()->basicLatin(),
+                    Set::strings()->unicode()->cyrillic(),
+                    Set::strings()->unicode()->arabic(),
+                )->map(\rawurlencode(...)),
+            )
             ->filter(static fn($value) => (bool) \preg_match('~\S+~', $value))
             ->exclude(static fn($value) => \str_contains($value, '//'))
-            ->exclude(static fn($value) => \str_starts_with($value, '\\'));
+            ->exclude(static fn($value) => \str_contains($value, '\\@'))
+            ->exclude(static fn($value) => \str_starts_with($value, '\\'))
+            ->map(static fn($value) => \trim($value, ' '))
+            ->exclude(static fn($value) => $value === '')
+            ->filter(static fn($value) => Url::attempt($value)->match(
+                static fn() => true,
+                static fn() => false,
+            ));
     }
 }