+++ tests/UrlTest.php
@@ -88,6 +88,14 @@ class UrlTest extends TestCase
         $this->assertSame($fragment, $url->fragment()->toString());
     }
 
+    public function testParseSchemeLessUrl()
+    {
+        $this->assertSame(
+            '//مثال.إختبار/',
+            Url::of('//مثال.إختبار/')->toString(),
+        );
+    }
+
     #[DataProvider('fromString')]
     #[DataProvider('parseable')]
     public function testValidStringsReturnAnUrl($string)
@@ -268,7 +276,7 @@ class UrlTest extends TestCase
     {
         $url = Url::of('//example.com');
 
-        $this->assertSame('example.com/', $url->toString());
+        $this->assertSame('//example.com/', $url->toString());
     }
 
     public function testPathIsAbsolute()
@@ -315,6 +323,50 @@ class UrlTest extends TestCase
                 ->unwrap()
                 ->toString(),
         );
+    }
+
+    public function testSchemeLessUrls(): BlackBox\Proof
+    {
+        return $this
+            ->forAll(Fixture::any())
+            ->prove(function($url) {
+                $this->assertStringStartsWith(
+                    '//',
+                    $url
+                        ->withScheme(Scheme::less())
+                        ->toString(),
+                );
+            });
+    }
+
+    public function testParsingIdempotencyOfPartiallyEncodedUrls()
+    {
+        $this->assertSame(
+            'http://h.tld/a%20b?x=%2e',
+            Url::of(Url::of('http://h.tld/a b?x=%2e')->toString())->toString(),
+        );
+    }
+
+    public function testAnyUrlFixtureCanBeParsed(): BlackBox\Proof
+    {
+        return $this
+            ->forAll(Fixture::any())
+            ->prove(function($url) {
+                $this->assert()->not()->throws(
+                    static fn() => Url::of($url->toString()),
+                );
+            });
+    }
+
+    public function testAnyUrlsCanBeResolved(): BlackBox\Proof
+    {
+        return $this
+            ->forAll(Fixture::any(), Fixture::any())
+            ->prove(function($a, $b) {
+                $this->assert()->not()->throws(
+                    static fn() => $a->resolve($b)->unwrap(),
+                );
+            });
     }
 
     public static function cases(): array
@@ -612,13 +664,14 @@ class UrlTest extends TestCase
                 'with=query',
                 'and-fragment',
             ],
+            ['//مثال.إختبار/', '', '', '', 'مثال.إختبار', '', '/', '', ''],
         ];
     }
 
     public static function resolvable(): array
     {
         return [
-            ['//example.com', 'foo', 'example.com/foo'],
+            ['//example.com', 'foo', '//example.com/foo'],
             ['http://example.com', 'foo', 'http://example.com/foo'],
             ['http://example.com/bar', 'http://example.com/foo', 'http://example.com/foo'],
             ['http://xn--example.com/foo/baz', './bar', 'http://xn--example.com/foo/bar'],